v0.0.26;

+++ fango/adapters/pydantic.py
@@ -7,6 +7,7 @@ from django.db.models import Model
 from django.db.models.fields.related import (
     ForeignKey,
     ManyToManyField,
+    ManyToManyRel,
     ManyToOneRel,
     OneToOneField,
     OneToOneRel,
@@ -17,11 +18,12 @@ from pydantic import BaseModel
 from fango.adapters.types import PK
 from fango.log import log_params
 from fango.schemas import CRUDAdapter, LinkAdapter
+from fango.utils import get_model_field_safe
 
 __all__ = ["PydanticAdapter"]
 
 ForwardRel = ForeignKey | OneToOneField | OneToOneRel
-MultipleRel = ManyToOneRel | ManyToManyField
+MultipleRel = ManyToOneRel | ManyToManyField | ManyToManyRel
 
 
 class AdapterError(Exception):
@@ -99,7 +101,6 @@ def _get_or_create_instance(
                 relation[remote_field.name] = remote_instance
             else:
                 relation[remote_field.name] = None
-
         else:
             raise AdapterError
 
@@ -123,7 +124,7 @@ def _get_handlers(instance: Model, model: type[Model], schema_instance: BaseMode
         if isinstance(getattr(model, key), property):
             continue
 
-        field = model._meta.get_field(key)
+        field = get_model_field_safe(model, key)
         args = (instance, field, key, value)
 
         if isinstance(field, OneToOneRel):
@@ -158,7 +159,7 @@ def _handle_multiple_relation(instance: Model, field: MultipleRel, key: str, val
     """
     relation_set = getattr(instance, key, None)
 
-    if value is None or value is []:
+    if value is None or value == []:
         if relation_set:
             try:
                 relation_set.clear()
@@ -169,20 +170,21 @@ def _handle_multiple_relation(instance: Model, field: MultipleRel, key: str, val
         _handle_crud_adapter(instance, field, key, value)
 
     elif isinstance(value, LinkAdapter):
-        _handle_link_adapter(instance, field, key, value)
+        _handle_link_adapter(instance, key, value)
 
     elif isinstance(value, list):
         data = []
         for item in value:
             data.append(_get_or_create_relation(instance, field, key, item))
 
-        try:
-            relation_set.set(data)
-        except AttributeError:
-            relation_set.all().delete()
-            for item in data:
-                item.clean()
-                item.save()
+        if relation_set:
+            try:
+                relation_set.set(data)
+            except AttributeError:
+                relation_set.all().delete()
+                for item in data:
+                    item.clean()
+                    item.save()
     else:
         raise AdapterError
 
@@ -260,7 +262,7 @@ def _handle_crud_adapter(instance: Model, field: MultipleRel, key: str, value: A
 
 
 @log_params("PoetryAdapter")
-def _handle_link_adapter(instance: Model, field: ForwardRel | MultipleRel, key: str, value: Any) -> None:
+def _handle_link_adapter(instance: Model, key: str, value: Any) -> None:
     """
     Manages Link adapter.
